Separate out CSV serialization & deserialization
